% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_schema}
\alias{get_schema}
\title{Get model schema}
\usage{
get_schema(model)
}
\arguments{
\item{model}{Model instance or constructor}
}
\value{
Schema definition
}
\description{
Get model schema
}
\examples{
User <- define_model(name = "character", age = "numeric")
user <- User(name = "John", age = 30)
get_schema(user)
get_schema(User)
}
