% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed_function.R
\name{get_signature}
\alias{get_signature}
\title{Get function signature}
\usage{
get_signature(fn)
}
\arguments{
\item{fn}{Typed function}
}
\value{
Signature information or NULL
}
\description{
Get function signature
}
\examples{
f <- typed_function(
  function(x, y) x + y,
  arg_specs = c(x = "numeric", y = "numeric"),
  return_spec = "numeric"
)
get_signature(f)
}
