% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_check.R
\name{is_one_of}
\alias{is_one_of}
\title{Check multiple type constraints}
\usage{
is_one_of(value, types)
}
\arguments{
\item{value}{The value to check}

\item{types}{Vector of allowed types}
}
\value{
logical
}
\description{
Check multiple type constraints
}
\examples{
is_one_of(5, c("numeric", "character"))
is_one_of("hello", c("numeric", "character"))
}
