% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_spec.R
\name{is_type_spec}
\alias{is_type_spec}
\title{Test whether an object is a type_spec}
\usage{
is_type_spec(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from class \code{type_spec}, otherwise \code{FALSE}.
}
\description{
Test whether an object is a type_spec
}
\examples{
is_type_spec(t_union("numeric", "character"))
is_type_spec("numeric")
}
