% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{list_of}
\alias{list_of}
\title{Validate list structure}
\usage{
list_of(element_type, min_length = 0, max_length = Inf)
}
\arguments{
\item{element_type}{Type of list elements}

\item{min_length}{Minimum list length}

\item{max_length}{Maximum list length}
}
\value{
validator function
}
\description{
Validate list structure
}
\examples{
num_list_validator <- list_of(element_type = "numeric", min_length = 1)
num_list_validator(list(1, 2, 3)) # TRUE
num_list_validator(list("a", "b")) # FALSE
}
