% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{print.typed_model}
\alias{print.typed_model}
\title{Print method for typed models}
\usage{
\method{print}{typed_model}(x, ...)
}
\arguments{
\item{x}{Model instance}

\item{...}{Additional arguments}
}
\description{
Print method for typed models
}
