% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{string_pattern}
\alias{string_pattern}
\title{String pattern validator (regex)}
\usage{
string_pattern(pattern, ignore_case = FALSE)
}
\arguments{
\item{pattern}{Regular expression pattern}

\item{ignore_case}{Ignore case in matching}
}
\value{
validator function
}
\description{
String pattern validator (regex)
}
\examples{
email_validator <- string_pattern("^[a-zA-Z0-9.+-]+@[a-zA-Z0-9.-]+$")
email_validator("user@example.com") # TRUE
email_validator("invalid-email") # FALSE
}
