% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_spec.R
\name{t_enum}
\alias{t_enum}
\title{Enumerated set of allowed values}
\usage{
t_enum(values)
}
\arguments{
\item{values}{Atomic vector of allowed values.}
}
\value{
A \code{type_spec} of kind \code{"enum"}.
}
\description{
Matches when the value is \code{\%in\%} the allowed set.
}
\examples{
spec <- t_enum(c("admin", "user", "guest"))
is_type("admin", spec)
is_type("root", spec)
}
