% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_spec.R
\name{t_nullable}
\alias{t_nullable}
\title{Allow \code{NULL} in addition to an inner type}
\usage{
t_nullable(type)
}
\arguments{
\item{type}{Inner type specification.}
}
\value{
A \code{type_spec} of kind \code{"nullable"}.
}
\description{
Allow \code{NULL} in addition to an inner type
}
\examples{
maybe_int <- t_nullable("integer")
is_type(NULL, maybe_int)
is_type(1L, maybe_int)
is_type("hi", maybe_int)
}
\seealso{
Other type specifications: 
\code{\link{is_type_spec}()},
\code{\link{t_enum}()},
\code{\link{t_list_of}()},
\code{\link{t_model}()},
\code{\link{t_predicate}()},
\code{\link{t_union}()},
\code{\link{t_vector_of}()},
\code{\link{type_spec}}
}
\concept{type specifications}
