% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed_function.R
\name{typed_method}
\alias{typed_method}
\title{Create method validator for R6 or S3 classes}
\usage{
typed_method(class_name, method_name, arg_types = list(), return_type = NULL)
}
\arguments{
\item{class_name}{Name of the class}

\item{method_name}{Name of the method}

\item{arg_types}{Argument types}

\item{return_type}{Return type}
}
\value{
Typed method wrapper
}
\description{
Create method validator for R6 or S3 classes
}
