% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{validate_model}
\alias{validate_model}
\title{Validate model instance}
\usage{
validate_model(instance)
}
\arguments{
\item{instance}{Model instance}
}
\value{
list with valid (logical) and errors (character vector)
}
\description{
Validate model instance
}
\examples{
User <- define_model(name = "character", age = "numeric")
user <- User(name = "John", age = 30, .validate_instance = FALSE)
validate_model(user)
}
