% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{vector_length}
\alias{vector_length}
\title{Vector length validator}
\usage{
vector_length(min_len = 0, max_len = Inf, exact_len = NULL)
}
\arguments{
\item{min_len}{Minimum length}

\item{max_len}{Maximum length}

\item{exact_len}{Exact length required}
}
\value{
validator function
}
\description{
Vector length validator
}
\examples{
pair_validator <- vector_length(exact_len = 2)
pair_validator(c(1, 2)) # TRUE
pair_validator(c(1, 2, 3)) # FALSE
}
