% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed_function.R
\name{with_signature}
\alias{with_signature}
\title{Apply type signature to function}
\usage{
with_signature(fn, sig)
}
\arguments{
\item{fn}{Function to type}

\item{sig}{Signature object from signature()}
}
\value{
Typed function
}
\description{
Apply type signature to function
}
\examples{
sig <- signature(x = "numeric", y = "numeric", .return = "numeric")
add <- with_signature(function(x, y) x + y, sig)
}
