% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacontract.R
\name{write_datacontract}
\alias{write_datacontract}
\title{Write a typethis schema to a Data Contract YAML file}
\usage{
write_datacontract(x, path, info = NULL, servers = NULL, ...)
}
\arguments{
\item{x}{See \code{\link[=to_datacontract]{to_datacontract()}}.}

\item{path}{Destination file path.}

\item{info, servers}{See \code{\link[=to_datacontract]{to_datacontract()}}.}

\item{...}{Forwarded to \code{\link[=to_datacontract]{to_datacontract()}}.}
}
\value{
The contract list, invisibly.
}
\description{
Write a typethis schema to a Data Contract YAML file
}
