% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{combine_validators}
\alias{combine_validators}
\title{Custom validator combinator}
\usage{
combine_validators(..., all_of = TRUE)
}
\arguments{
\item{...}{Validator functions}

\item{all_of}{If TRUE, all validators must pass; if FALSE, any can pass}
}
\value{
Combined validator function
}
\description{
Custom validator combinator
}
\examples{
validator <- combine_validators(
  is_numeric <- function(x) is.numeric(x),
  is_positive <- function(x) all(x > 0),
  all_of = TRUE
)
validator(5) # TRUE
validator(-5) # FALSE
}
