% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacontract.R
\name{datacontract_cli_available}
\alias{datacontract_cli_available}
\title{Check whether the \code{datacontract} CLI is available on PATH}
\usage{
datacontract_cli_available()
}
\value{
\code{TRUE} if the binary is found, \code{FALSE} otherwise.
}
\description{
Check whether the \code{datacontract} CLI is available on PATH
}
\examples{
datacontract_cli_available()
}
\seealso{
Other Data Contract: 
\code{\link{datacontract}},
\code{\link{datacontract_export}()},
\code{\link{datacontract_lint}()},
\code{\link{datacontract_test}()},
\code{\link{from_datacontract}()},
\code{\link{read_datacontract}()},
\code{\link{to_datacontract}()},
\code{\link{write_datacontract}()}
}
\concept{Data Contract}
