% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{dataframe_spec}
\alias{dataframe_spec}
\title{Data frame validator}
\usage{
dataframe_spec(required_cols = NULL, min_rows = 0, max_rows = Inf)
}
\arguments{
\item{required_cols}{Required column names}

\item{min_rows}{Minimum number of rows}

\item{max_rows}{Maximum number of rows}
}
\value{
validator function
}
\description{
Data frame validator
}
\examples{
df_validator <- dataframe_spec(
  required_cols = c("id", "name"),
  min_rows = 1
)
df <- data.frame(id = 1:3, name = c("A", "B", "C"))
df_validator(df) # TRUE
}
