% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{enum_validator}
\alias{enum_validator}
\title{Create enum validator}
\usage{
enum_validator(allowed_values)
}
\arguments{
\item{allowed_values}{Vector of allowed values}
}
\value{
validator function
}
\description{
Create enum validator
}
\examples{
status_validator <- enum_validator(c("active", "inactive", "pending"))
status_validator("active") # TRUE
status_validator("deleted") # FALSE
}
