% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_spec.R
\name{is_type_spec}
\alias{is_type_spec}
\title{Test whether an object is a type spec}
\usage{
is_type_spec(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Returns \code{TRUE} if \code{x} was created by one of the \verb{t_*()} constructors.
}
\examples{
is_type_spec(t_union("numeric", "character"))
is_type_spec("numeric")
}
\seealso{
Other type specifications:
\code{\link[=t_enum]{t_enum()}},
\code{\link[=t_list_of]{t_list_of()}},
\code{\link[=t_model]{t_model()}},
\code{\link[=t_nullable]{t_nullable()}},
\code{\link[=t_predicate]{t_predicate()}},
\code{\link[=t_union]{t_union()}},
\code{\link[=t_vector_of]{t_vector_of()}},
\code{\link{type_spec}}
}
\concept{type specifications}
