% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed_function.R
\name{is_typed}
\alias{is_typed}
\title{Check if function is typed}
\usage{
is_typed(fn)
}
\arguments{
\item{fn}{Function to check}
}
\value{
logical
}
\description{
Check if function is typed
}
\examples{
f1 <- function(x) x + 1
f2 <- typed_function(f1, arg_specs = c(x = "numeric"))
is_typed(f1) # FALSE
is_typed(f2) # TRUE
}
