% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{nullable}
\alias{nullable}
\title{Nullable type validator}
\usage{
nullable(validator)
}
\arguments{
\item{validator}{Base validator function}
}
\value{
Validator that also accepts NULL
}
\description{
Nullable type validator
}
\examples{
optional_num <- nullable(function(x) is.numeric(x))
optional_num(5) # TRUE
optional_num(NULL) # TRUE
}
