% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{numeric_range}
\alias{numeric_range}
\title{Numeric range validator}
\usage{
numeric_range(
  min = -Inf,
  max = Inf,
  exclusive_min = FALSE,
  exclusive_max = FALSE
)
}
\arguments{
\item{min}{Minimum value (inclusive)}

\item{max}{Maximum value (inclusive)}

\item{exclusive_min}{Minimum is exclusive}

\item{exclusive_max}{Maximum is exclusive}
}
\value{
validator function
}
\description{
Numeric range validator
}
\examples{
age_validator <- numeric_range(min = 0, max = 120)
age_validator(25) # TRUE
age_validator(150) # FALSE
}
