% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{string_length}
\alias{string_length}
\title{String length validator}
\usage{
string_length(min_length = 0, max_length = Inf)
}
\arguments{
\item{min_length}{Minimum string length}

\item{max_length}{Maximum string length}
}
\value{
validator function
}
\description{
String length validator
}
\examples{
name_validator <- string_length(min_length = 1, max_length = 50)
name_validator("John") # TRUE
name_validator("") # FALSE
}
