% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{update_model}
\alias{update_model}
\title{Update model fields}
\usage{
update_model(model, ..., .validate = TRUE)
}
\arguments{
\item{model}{Model instance}

\item{...}{Fields to update}

\item{.validate}{Validate after update}
}
\value{
Updated model instance
}
\description{
Update model fields
}
\examples{
User <- define_model(name = "character", age = "numeric")
user <- User(name = "John", age = 30)
user <- update_model(user, age = 31)
}
