% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{validate_field_value}
\alias{validate_field_value}
\title{Validate a field value against its definition}
\usage{
validate_field_value(fname, value, field_def, class_name = "model")
}
\arguments{
\item{fname}{Field name}

\item{value}{Field value}

\item{field_def}{Field definition list}

\item{class_name}{Model class name for error messages}
}
\description{
Validate a field value against its definition
}
\keyword{internal}
