% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed_function.R
\name{with_signature}
\alias{with_signature}
\title{Apply a signature to a function}
\usage{
with_signature(fn, sig)
}
\arguments{
\item{fn}{Function to wrap.}

\item{sig}{A \code{type_signature} object from \code{\link[=signature]{signature()}}.}
}
\value{
A typed function.
}
\description{
Equivalent to passing \code{sig$args} and \code{sig$return} to \code{\link[=typed_function]{typed_function()}}.
}
\examples{
sig <- signature(x = "numeric", y = "numeric", .return = "numeric")
multiply <- with_signature(function(x, y) x * y, sig)
multiply(5, 3)
}
\seealso{
Other typed functions: 
\code{\link{as_typed}()},
\code{\link{as_typed_env}()},
\code{\link{enable_for_package}()},
\code{\link{get_signature}()},
\code{\link{infer_specs}()},
\code{\link{is_typed}()},
\code{\link{signature}},
\code{\link{typed_function}()},
\code{\link{typed_method}()},
\code{\link{types}()},
\code{\link{validate_call}()}
}
\concept{typed functions}
